Conflicts:
	com.openexchange.contact.storage.rdb/src/com/openexchange/contact/storage/rdb/fields/QueryFields.java

+++ com.openexchange.index/src/com/openexchange/index/IndexResult.java
@@ -64,7 +64,7 @@ public interface IndexResult<V> {
 	 * 
 	 * @return The total number of results.
 	 */
-	long getNumFound();
+	int getNumFound();
 	
 	/**
 	 * Gets the requested results as specified by {@link QueryParameters}'s paging.
